Moved Exception classes in exceptions module
* This way I can mock out all of GraphAPI class
